// Frontend/src/generated/extensions/model/candle.ts
/**
 * Trade REST API -- our extensions
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Candle { 
    date: string;
    open: number;
    high: number;
    low: number;
    close: number;
    /**
     * Absent on a day Fauxnance filled in.
     */
    volume?: number | null;
}